Issue #4043: Split and Organize Checkstyle inputs by Test for OuterTypeNumber
